Fix / Remove unnecessary modules and simplify imports
