--- src/config/__init__.py
@@ -1,1 +1,0 @@
-from ._Configuration import _Configuration as Configuration
